FIX SSAO basic logic - storing viewPos in gbuffer   + somewhat more acccurate than deducing from zbuffer   + also viewZ was already stored - rearrange shader units

+++ src/render/FrameBufferAttachment.cpp
@@ -420,7 +420,7 @@ namespace render {
         FrameBufferAttachment spec;
         spec.type = FrameBufferAttachmentType::texture;
         // NOTE KI GL_RGB16F is *NOT* enough precision; causes very block light/shadows
-        spec.internalFormat = GL_RGB32F;
+        spec.internalFormat = GL_RGB16F;
         spec.attachment = attachment;
         spec.useDrawBuffer = true;
         spec.clearMask = GL_COLOR_BUFFER_BIT;
